refactor: extract schema primitives
